Add has / hasNot shortcut methods
This has(Class<?> clazz) method is in perticular very useful when retrieving large lists of elements.
The PolymorphicTypeResolver would invoke a getProperty call for each element of
the list in order to verify the type. It is up to 40x faster to add this type check to
the gremlin pipeline using e.g. has(Person.class) and toListExplicit(Person.class)
instead of just toList(Person.class).